Add commented code to "fix" initialization.
Sometimes initializing with old data makes it not load correctly the
libraries. To solve this, new information about the libraries that are
available are downloaded.

As usually looking for this takes me 10 minutes (because I don't learn
how to look more efficiently for it), I leave it commented as a
reminder.